added more pages... for the 'next' button. (it's wrong)
It's wrong to manually do what could be done with code.
How to iterate over array fetched from API, to later display it on the page?

I tried, but my code was shit (and didn't work).

there is probably a simple and elegant solution for this, which I'm missing, because I don't know enough react and javascript